requirements.txt
factory_boy
lockfile
mock
pytest
pytest-django
python-dateutil
